requirements.txt
feedparser>=6.0.10
PyYAML>=6.0
anthropic>=0.40.0
httpx>=0.27.0
python-dotenv>=1.0.0
flask>=3.0.0
# On Ubuntu/Debian: sudo apt install python3-feedparser python3-yaml
# Everything else: pip install anthropic httpx python-dotenv flask
